Change data type of mythicPlusScore to SMALLINT

+++ models/WoWCharacters.js
@@ -21,7 +21,7 @@ module.exports = class WoWCharacters extends Model {
         soloShuffleSpec3Rating: { type: DataTypes.INTEGER, allowNull: true },
         soloShuffleSpec4Rating: { type: DataTypes.INTEGER, allowNull: true },
         specialization: { type: DataTypes.STRING, allowNull: true },
-        mythicPlusScore: { type: DataTypes.FLOAT, allowNull: true },
+        mythicPlusScore: { type: DataTypes.SMALLINT, allowNull: true },
       },
       {
         tableName: "WoWCharacters",
